requirements.txt
# Python dependencies for NASA SWOT Monitoring System

# Data processing
pandas>=1.5.0
numpy>=1.23.0

# NASA Earthdata access
requests>=2.28.0
python-dotenv>=0.19.0

# Statistical analysis
scipy>=1.9.0

# Visualization
matplotlib>=3.6.0

# Configuration
PyYAML>=6.0
